Test Generalization
Add test that creates a generalization on two diagrams.
Both represent the same generalization.
